Update spectrometer.py
The channels A to F belong to the AS72653, and the channels R to W belong to the AS72651. The function reorderData sort the channels starting from A, which means that the first sensor to read is the AS72653, not the AS72651.

I created a bar chart with the light wavelength, and I could make the correct plot only with that change.